back_end/requirements.txt
# main.ipynb — KaggleDatasetAdapter.PANDAS
kagglehub[pandas-datasets]

# Notebook (progress bars / widgets)
jupyter
ipywidgets

# Optional: load KAGGLE_* from .env
python-dotenv

# FSQ OS Places download/filter pipeline
duckdb
huggingface_hub
pyarrow

# Date Night filtering pipeline
httpx
numpy
pandas
pyyaml
crawl4ai
